+++ flowman-core/src/main/scala/com/dimajix/flowman/config/FlowmanConf.scala
@@ -32,8 +32,9 @@ import com.dimajix.flowman.execution.SimpleExecutor
 import com.dimajix.flowman.execution.DependencyScheduler
 import com.dimajix.flowman.execution.Scheduler
 import com.dimajix.flowman.model.VerifyPolicy
-import com.dimajix.flowman.transforms.ColumnMismatchStrategy
-import com.dimajix.flowman.transforms.TypeMismatchStrategy
+import com.dimajix.flowman.transforms.ColumnMismatchPolicy
+import com.dimajix.flowman.transforms.CharVarcharPolicy
+import com.dimajix.flowman.transforms.TypeMismatchPolicy
 import com.dimajix.spark.features
 
 
@@ -120,22 +121,30 @@ object FlowmanConf {
         .stringConf
         .createWithDefault(MigrationStrategy.ALTER.toString)
 
-    val DEFAULT_RELATION_INPUT_COLUMN_MISMATCH_STRATEGY = buildConf("flowman.default.relation.input.columnMismatchPolicy")
+    val DEFAULT_RELATION_INPUT_COLUMN_MISMATCH_POLICY = buildConf("flowman.default.relation.input.columnMismatchPolicy")
         .doc("Default strategy to use on schema column mismatch while reading relations. Can be 'ignore', 'error', 'add_columns_or_ignore', 'add_columns_or_error', 'remove_columns_or_ignore', 'remove_columns_or_error', 'add_remove_columns'")
         .stringConf
-        .createWithDefault(ColumnMismatchStrategy.IGNORE.toString)
-    val DEFAULT_RELATION_INPUT_TYPE_MISMATCH_STRATEGY = buildConf("flowman.default.relation.input.typeMismatchPolicy")
+        .createWithDefault(ColumnMismatchPolicy.IGNORE.toString)
+    val DEFAULT_RELATION_INPUT_TYPE_MISMATCH_POLICY = buildConf("flowman.default.relation.input.typeMismatchPolicy")
         .doc("Default strategy to use on schema type mismatch while reading relations. Can be 'ignore', 'error', 'cast_compatible_or_ignore', 'cast_compatible_or_error', 'cast_always'")
         .stringConf
-        .createWithDefault(TypeMismatchStrategy.IGNORE.toString)
-    val DEFAULT_RELATION_OUTPUT_COLUMN_MISMATCH_STRATEGY = buildConf("flowman.default.relation.output.columnMismatchPolicy")
+        .createWithDefault(TypeMismatchPolicy.IGNORE.toString)
+    val DEFAULT_RELATION_INPUT_CHAR_VARCHAR_POLICY = buildConf("flowman.default.relation.input.charVarcharPolicy")
+        .doc("Default strategy to use when reading CHAR(n)/VARCHAR(n) data types. Can be 'ignore', 'pad', 'truncate' or 'pad_and_truncate'")
+        .stringConf
+        .createWithDefault(CharVarcharPolicy.PAD_AND_TRUNCATE.toString)
+    val DEFAULT_RELATION_OUTPUT_COLUMN_MISMATCH_POLICY = buildConf("flowman.default.relation.output.columnMismatchPolicy")
         .doc("Default strategy to use on schema column mismatch while reading relations. Can be 'ignore', 'error', 'add_columns_or_ignore', 'add_columns_or_error', 'remove_columns_or_ignore', 'remove_columns_or_error', 'add_remove_columns'")
         .stringConf
-        .createWithDefault(ColumnMismatchStrategy.ADD_REMOVE_COLUMNS.toString)
-    val DEFAULT_RELATION_OUTPUT_TYPE_MISMATCH_STRATEGY = buildConf("flowman.default.relation.output.typeMismatchPolicy")
+        .createWithDefault(ColumnMismatchPolicy.ADD_REMOVE_COLUMNS.toString)
+    val DEFAULT_RELATION_OUTPUT_TYPE_MISMATCH_POLICY = buildConf("flowman.default.relation.output.typeMismatchPolicy")
         .doc("Default strategy to use on schema type mismatch while reading relations. Can be 'ignore', 'error', 'cast_compatible_or_ignore', 'cast_compatible_or_error', 'cast_always'")
         .stringConf
-        .createWithDefault(TypeMismatchStrategy.CAST_ALWAYS.toString)
+        .createWithDefault(TypeMismatchPolicy.CAST_ALWAYS.toString)
+    val DEFAULT_RELATION_OUTPUT_CHAR_VARCHAR_POLICY = buildConf("flowman.default.relation.output.charVarcharPolicy")
+        .doc("Default strategy to use when writing CHAR(n)/VARCHAR(n) data types. Can be 'ignore', 'pad', 'truncate' or 'pad_and_truncate'")
+        .stringConf
+        .createWithDefault(CharVarcharPolicy.PAD_AND_TRUNCATE.toString)
 
     val DEFAULT_TARGET_VERIFY_POLICY = buildConf("flowman.default.target.verifyPolicy")
         .doc("Policy for verifying a target. Accepted verify policies are 'empty_as_success', 'empty_as_failure' and 'empty_as_success_with_errors'.")