+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * github-226: Upgrade to Spark 3.2.2
 * github-227: [BUG] Flowman should not fail with field names containing "-", "/" etc
+* github-228: Padding and truncation of CHAR(n)/VARCHAR(n) should be configurable
 
 
 # Version 0.26.0 - 2022-07-27